pyproject.toml
[project]
name = "saas-boilerplate"
version = "0.1.0"
description = "FastAPI SaaS Boilerplate — Production Ready"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # --- Web Framework ---
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",

    # --- Database (SQLModel = SQLAlchemy 2.0 + Pydantic) ---
    "sqlmodel>=0.0.19",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",          # PostgreSQL async driver
    "alembic>=1.13.0",          # Migrations

    # --- Auth ---
    "python-jose[cryptography]>=3.3.0",  # JWT tokens
    "passlib[bcrypt]>=1.7.4",            # Password hashing

    # --- Config & Env ---
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # --- Cache & Sessions ---
    "redis>=5.0.0",

    # --- HTTP Client ---
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

# ── uv dev dependencies (alternative to optional-dependencies) ──
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

# ── Linter (Ruff) ──
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]  # Allow function calls in default args (FastAPI Depends)

# ── Tests ──
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

# ── Type Checking ──
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
